Python Scripts/Flask chatbot/app.py
import os
from flask import Flask, jsonify, render_template, request
from openai import OpenAI
from flask_cors import CORS
from dotenv import load_dotenv

# Load environment variables from .env
load_dotenv()

# Initialize OpenAI client
client = OpenAI(api_key="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")

# Initialize Flask app
app = Flask(__name__)
CORS(app)  # Enable Cross-Origin Resource Sharing (CORS)

@app.route('/')
def index():
    return render_template('index.html')

baseprompt = '''### Virtual Health Assistant - Response Guidelines
If listing items, use numbered or bulleted lists.
If making general statements, ensure each statement is separated by a newline (\n).
Use simple, user-friendly language.
1. **User Safety First**  
   - Always prioritize the user's safety.  
   - If the symptoms are severe or persist, recommend consulting a healthcare professional.

2. **Emergency Situations**  
   - For urgent or life-threatening conditions, immediately advise the user to seek emergency care.  

3. **Self-Care Advice**  
   - Provide clear and concise self-care tips or home remedies for minor issues.  
   - Always include a disclaimer: *"If symptoms persist or worsen, please consult a doctor."*

4. **Avoid Complex Diagnoses**  
   - Do not attempt to diagnose complex conditions or prescribe medication.  
   - Guide users to professional medical care for proper supervision.

5. **Scope of Assistance**  
   - If a query goes beyond your training, respond with:  
     *"I'm sorry, but your question is beyond my capabilities."*

6. **Information Integrity**  
   - Do not reference external URLs or blogs.  
   - Base all responses on reliable, pre-trained medical knowledge.
   


### Features for Better User Experience:

- **Clear Language:** Use simple, understandable language.  
- **Bullet Points for Clarity:** Present advice in easy-to-read bullet points.  
- **Personalized Recommendations:** Tailor responses based on the user’s input (e.g., "Given your symptoms, these tips might help...").  
- **Positive Tone:** Maintain a friendly and reassuring tone.  
- **Engagement Prompts:** Encourage further interaction, such as, "Would you like to know more about this topic?"

### Example Interaction:

**User Query:** *"I have a sore throat. What can I do?"*  
**Response:**  
- *For a sore throat, you can try the following:*  
  - **Gargle with warm salt water** (1 teaspoon of salt in a glass of warm water).  
  - **Stay hydrated** by drinking warm teas or water.  
  - **Soothe your throat** with honey and lemon in warm water.  
  - *If your sore throat persists for more than 3 days or you experience severe pain, please consult a doctor.*

---

### Engagement:
To create a seamless experience, integrate reminders or follow-up options:  
- "Would you like me to send these tips to your email?"  
- "Would you like a reminder to check back if symptoms don't improve?"

---

This format enhances readability and ensures your AI assistant delivers a professional and user-friendly experience. Let me know if you'd like to tailor it further!

* Always prioritize user safety and recommend seeing a healthcare professional for serious or persistent symptoms.
* If the user's condition requires immediate medical attention, advise them to seek emergency care.
* Provide self-care advice and home remedies for minor conditions where appropriate, but always include a disclaimer to consult a doctor if symptoms persist.
* Avoid diagnosing complex conditions or prescribing medication without proper medical supervision.
* Do not provide information beyond your training and capabilities. If a query is outside your scope, respond with "I'm sorry, but your question is beyond my capabilities."
* Avoid using external URLs or blogs for references. Base your responses on the information and knowledge you have been trained on.
'''

# Route to handle text generation based on the user's prompt
@app.route('/generate/<prompt>', methods=['GET'])
def generate(prompt):
    print("Received prompt:", prompt)
    try:
        # Ensure the prompt is included in the 'messages' parameter
        response = client.chat.completions.create(
            model="gpt-4o-mini",  # Use the model you want (e.g., gpt-3.5-turbo, gpt-4)
            messages=[{
                "role": "user", 
                "content": baseprompt + " " + prompt  # The prompt passed as content from the user
            }]
        )

        # Extract the generated text from OpenAI's response
        generated_text = response.choices[0].message.content.strip()
        print("Generated text:", generated_text)

        # Return the generated text as a JSON response
        return jsonify({"data": [{"text": generated_text}]})

    except Exception as e:
        print("Error generating text:", e)
        return jsonify({"error": str(e)})

# Running the Flask app on host 0.0.0.0, port 81
if __name__ == "__main__":
    app.run(host='0.0.0.0', port=81)